🌱 Split oversized codeGenerator.templates.ts into focused modules
Split 1326-line file into category-based modules:
- templates/clusterCards.ts (272 lines)
- templates/infrastructureCards.ts (355 lines)
- templates/cicdCards.ts (393 lines)
- templates/cardRenderFunction.ts (37 lines)
- templates/shared.ts (54 lines)

The main file now re-exports everything for backward compatibility.
All files are now under 500 lines as required by Auto-QA.

Part of #19750
